feat: add SupplierArrivalLineDetailQuantityCard component
